requirements.txt
python-telegram-bot
configparser